+++ web/static/css/main.css
@@ -19,7 +19,7 @@ a.button{display:inline-block; padding: 10px 20px; text-decoration:none; color:#
   background:rgba(0,0,0,0.3);
 }
 
-.modal-con, .modal-con2, .modal-con3, .modal-con4 {
+.modal-con, .modal-con2, .modal-con3, .modal-con4{
   display:none;
   position:fixed;
   top:50%; left:50%;
@@ -41,6 +41,7 @@ a.button{display:inline-block; padding: 10px 20px; text-decoration:none; color:#
   line-height:1.3;
   padding: 30px;
 }
+
 
 .modal-con .close, .modal-con2 .close, .modal-con3 .close,  .modal-con4 .close {
   display:block;
@@ -65,6 +66,31 @@ a.button{display:inline-block; padding: 10px 20px; text-decoration:none; color:#
   padding : 5px;
   margin: 20px;
   height: 50px;
+}
+.modal-con4 .con .container{
+  width: 600px;
+  height: 300px;
+  overflow: auto;
+}
+.modal-con4 .con table{
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.modal-con4 .con th, td{
+  padding: 10px;
+  border-bottom: 1px solid #444444 !important;
+}
+
+
+.modal-con4 .con .container::-webkit-scrollbar{
+  width:10px;
+}
+.modal-con4 .con .container::-webkit-scrollbar-thumb{
+  background:rgba(0,0,0,0.3);
+}
+.modal-con4 .con .container::-webkit-scrollbar-track{
+  background-color:whitesmoke;
 }
 
 .login input {
@@ -116,7 +142,7 @@ a.button{display:inline-block; padding: 10px 20px; text-decoration:none; color:#
 .areaName {
   position: relative;
   top: 80px;
-  margin: 0 !important;
+  margin: 0
 }
 
 .bar {
@@ -128,7 +154,7 @@ a.button{display:inline-block; padding: 10px 20px; text-decoration:none; color:#
   height: 180px;
 }
 
-table {
+.text-center .parking table {
   margin-left:auto;
   margin-right:auto;
   border-collapse: separate;
@@ -136,24 +162,24 @@ table {
   margin-bottom: 10px;
 }
 
-thead, tbody, tfoot, tr, td, th {
-  border: none !important;
+.text-center .parking thead, tbody, tfoot, tr, td, th {
+  border: none;
 }
 
-th, td {
+.text-center .parking th, td {
   padding: 10px;
   border-radius: 5px;
 }
 
-td {
+.text-center .parking td {
   width: 100px;
   height: 100px;
 }
 
-td#blocked {
+.text-center .parking td#blocked {
   background-color: rgb(214, 62, 62);
 }
 
-td#notBlocked {
+.text-center .parking td#notBlocked {
   background-color: rgb(109, 167, 109);
 }
